Fix communication between Tuner and Web interface

+++ Cosette.Tuner.Web/ViewModels/GenerationViewModel.cs
@@ -11,7 +11,7 @@ namespace Cosette.Tuner.Web.ViewModels
 
         public DateTime CreationTimeUtc { get; set; }
         public double ElapsedTime { get; set; }
-        public int BestFitness { get; set; }
+        public double BestFitness { get; set; }
 
         public List<GeneViewModel> BestGenes { get; set; }
     }
